wip: Attempt to fix StackFrame::IteratePc() in lang/node
This is an attempt to fix the addons tests which are still failing,
even after fixing the return address decoding for EXIT stack frames.

IteratePc should not blindly dereference the pc_address that is passed
to it, as the stored Program Counter could require "unmangling" on some
architectures, even before retguard.  At the very least,
__builtin_extract_return_addr should probably be applied.

However, even with this change, the addons tests are occasionally
failing when GC gets triggered during the test.